+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## In development
 
 - Fixes to issues discovered during deploy process [#3](https://github.com/nre-learning/antidote-ui-components/pull/3)
+- Add class to collection logo, fix l8n strings [#10](https://github.com/nre-learning/antidote-ui-components/pull/10)
 
 ## v0.4.0 - N/A